feat: Add a generalized file upload feature for all models, removing the file after successful database upload
