+++ src/certificateModule/routes/index.js
@@ -2,6 +2,5 @@ const express = require("express");
 const router = express.Router();
 
 router.use("/certificate", require("./certificate"));
-router.use("/upload", require("./uploadCertificate"));
 
 module.exports = router;
