+++ views/index.ejs
@@ -5,7 +5,7 @@
 </head>
 <body>
   <h1>Upload a CSV File</h1>
-  <form action="/certificateModule/upload" method="post" enctype="multipart/form-data">
+  <form action="/upload/announcement" method="post" enctype="multipart/form-data">
     <input type="file" name="csvFile">
     <input type="submit" value="Upload">
   </form>
